chore: fixing form css and renaming pages
