+++ app/tests/unit/pages/project-revision/[projectRevision]/create.test.tsx
@@ -2,35 +2,33 @@ import "@testing-library/jest-dom";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { ProjectRevisionCreate } from "pages/cif/project-revision/[projectRevision]/create";
+import { getProjectRevisionPageRoute } from "routes/pageRoutes";
 import PageTestingHelper from "tests/helpers/pageTestingHelper";
 
-import {
-  compliledcreateProjectRevisionQuery,
+import compliledcreateProjectRevisionQuery, {
   createProjectRevisionQuery,
 } from "__generated__/createProjectRevisionQuery.graphql";
-
 const defaultMockResolver = {
-  Project() {
+  ProjectRevision() {
     return {
-      pendingNewProjectRevision: "mock-revision-id",
+      id: "mock-id",
     };
   },
 };
-
 const pageTestingHelper = new PageTestingHelper<createProjectRevisionQuery>({
   pageComponent: ProjectRevisionCreate,
   compiledQuery: compliledcreateProjectRevisionQuery,
   defaultQueryResolver: defaultMockResolver,
   defaultQueryVariables: {
-    projectRevision: "test-cif-project-revision",
+    projectRevision: "mock-id",
   },
 });
 
 describe("The project amendments and revisions page", () => {
   beforeEach(() => {
     pageTestingHelper.reinit();
+    jest.restoreAllMocks();
   });
-
   it("renders an Amendment radio button", () => {
     pageTestingHelper.loadQuery();
     pageTestingHelper.renderPage();
@@ -53,16 +51,18 @@ describe("The project amendments and revisions page", () => {
     ).toBeInTheDocument();
   });
 
-  it("Triggers the projectRevisionCreateNewQuery and redirects when the user clicks the new revision button", () => {
+  it("Triggers the createProjectRevisionQuery and redirects when the user clicks the new revision button", () => {
     pageTestingHelper.loadQuery();
     pageTestingHelper.renderPage();
 
+    userEvent.click(screen.getByLabelText(/minor revision/i));
     userEvent.click(
       screen.getAllByRole("button", { name: /new revision/i })[0]
     );
 
-    const operation =
-      pageTestingHelper.environment.mock.getMostRecentOperation();
-    expect(operation.fragment.node.name).toBe("projectRevisionCreateNewQuery");
+    // pageTestingHelper.expectMutationToBeCalled("createProjectRevision", {});
+    expect(pageTestingHelper.router.push).toHaveBeenCalledWith(
+      getProjectRevisionPageRoute("mock-id")
+    );
   });
 });
